Add canonical_musicbrainz_data_release table
Same as canonical_musicbrainz_data with release name in combined lookup.
