+++ mbid_mapping/mapping/canonical_musicbrainz_data_base.py
@@ -1,12 +1,12 @@
 import re
 
 import psycopg2
+from psycopg2.errors import OperationalError
 from unidecode import unidecode
 
-from mapping.canonical_musicbrainz_data_base import CanonicalMusicBrainzDataBase
-from mapping.canonical_musicbrainz_data_release import CanonicalMusicBrainzDataRelease
 from mapping.utils import log
 from mapping.custom_sorts import create_custom_sort_tables
+from mapping.bulk_table import BulkInsertTable
 from mapping.canonical_recording_redirect import CanonicalRecordingRedirect
 from mapping.canonical_recording_release_redirect import CanonicalRecordingReleaseRedirect
 from mapping.canonical_release_redirect import CanonicalReleaseRedirect
@@ -14,14 +14,16 @@ from mapping.canonical_release import CanonicalRelease
 
 import config
 
+TEST_ARTIST_IDS = [1160983, 49627, 65, 21238]  # Gun'n'roses, beyoncé, portishead, Erik Satie
 
-class CanonicalMusicBrainzData(CanonicalMusicBrainzDataBase):
-    """
-        This class creates the MBID mapping tables without release name in the lookup.
-    """
 
-    def __init__(self, mb_conn, lb_conn=None, batch_size=None):
-        super().__init__("mapping.canonical_musicbrainz_data", mb_conn, lb_conn, batch_size)
+class CanonicalMusicBrainzDataBase(BulkInsertTable):
+    """
+        This class creates the MBID mapping tables.
+
+        For documentation on what each of the functions in this class does, please refer
+        to the BulkInsertTable docs.
+    """
 
     def get_create_table_columns(self):
         return [("id",                 "SERIAL"),
@@ -60,7 +62,7 @@ class CanonicalMusicBrainzData(CanonicalMusicBrainzDataBase):
                    ON m.id = t.medium
                  JOIN musicbrainz.release rl
                    ON rl.id = m.release
-                 JOIN mapping.canonical_musicbrainz_data_release_tmp rpr
+                 JOIN mapping.canonical_release_tmp rpr
                    ON rl.id = rpr.release
                  JOIN (SELECT artist_credit, array_agg(gid ORDER BY position) AS artist_mbids
                          FROM musicbrainz.artist_credit_name acn2
@@ -70,43 +72,38 @@ class CanonicalMusicBrainzData(CanonicalMusicBrainzDataBase):
                    ON acn.artist_credit = s.artist_credit
             LEFT JOIN musicbrainz.release_country rc
                    ON rc.release = rl.id
-               --- there is some bad data in MB for which the title is too large and exceeds the postgres indexing limits
-               --- therefore filter out such recordings before-hand, otherwise index creation may fail
-                WHERE length(concat(ac.name, r.name)) < 500      
              GROUP BY rpr.id, ac.id, s.artist_mbids, rl.gid, artist_credit_name, r.gid, r.name, release_name, year
              ORDER BY ac.id, rpr.id
         """)]
 
-    def get_post_process_queries(self):
-        return ["""
-            WITH all_recs AS (
-                SELECT *
-                     , row_number() OVER (PARTITION BY combined_lookup ORDER BY score) AS rnum
-                  FROM mapping.canonical_musicbrainz_data_tmp
-            ), deleted_recs AS (
-                DELETE
-                  FROM mapping.canonical_musicbrainz_data_tmp
-                 WHERE id IN (SELECT id FROM all_recs WHERE rnum > 1)
-             RETURNING recording_mbid, combined_lookup
-            )
-           INSERT INTO mapping.canonical_recording_redirect_tmp (recording_mbid, canonical_recording_mbid, canonical_release_mbid)
-                SELECT t1.recording_mbid
-                     , t2.recording_mbid AS canonical_recording
-                     , t2.release_mbid AS canonical_release
-                  FROM deleted_recs t1
-                  JOIN all_recs t2
-                    ON t1.combined_lookup = t2.combined_lookup
-                 WHERE t2.rnum = 1
-                 -- some recording mbids appear on multiple releases and the insert query inserts them once for
-                 -- for each appearance with the appropriate release mbid. the deletion criteria is combined_lookup
-                 -- which is unavailable in the insert sql query so we cannot easily apply a filter there itself.
-                 -- such rows  cleaned up in the deleted_recs with above so to above adding a redirect to the same
-                 -- recording as a canonical_recording, this condition.
-                   AND t1.recording_mbid != t2.recording_mbid;
-        """]
+    def get_index_names(self):
+        table = self.table_name.split(".")[-1]
+        return [
+            (f"{table}_idx_combined_lookup",              "combined_lookup", False),
+            (f"{table}_idx_artist_credit_recording_name_release_name", "artist_credit_name, recording_name, release_name", False),
+            (f"{table}_idx_recording_mbid", "recording_mbid", True)
+        ]
 
     def get_combined_lookup(self, row):
-        return unidecode(re.sub(r'[^\w]+', '', row['artist_credit_name'] + row['recording_name']).lower())
+        pass
+
+    def process_row(self, row):
+        combined_lookup = self.get_combined_lookup(row)
+        return {
+            self.table_name: [(
+                row["artist_credit_id"],
+                row["artist_mbids"],
+                row["artist_credit_name"],
+                row["release_mbid"],
+                row["release_name"],
+                row["recording_mbid"],
+                row["recording_name"],
+                combined_lookup,
+                row["score"],
+                row["year"]
+            )]
+        }
+
 
 
 def create_canonical_musicbrainz_data(use_lb_conn: bool):
@@ -131,12 +128,10 @@ def create_canonical_musicbrainz_data(use_lb_conn: bool):
         releases = CanonicalRelease(mb_conn)
         mapping = CanonicalMusicBrainzData(mb_conn, lb_conn)
         mapping.add_additional_bulk_table(can)
-        mapping_release = CanonicalMusicBrainzDataRelease(mb_conn, lb_conn)
 
         # Carry out the bulk of the work
         create_custom_sort_tables(mb_conn)
         releases.run(no_swap=True)
-        mapping_release.run(no_swap=True)
         mapping.run(no_swap=True)
         can_rec_rel.run(no_swap=True)
         can_rel.run(no_swap=True)
@@ -145,7 +140,6 @@ def create_canonical_musicbrainz_data(use_lb_conn: bool):
         log("canonical_musicbrainz_data: Swap into production")
         if lb_conn:
             releases.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
-            mapping_release.swap_into_production(no_swap_transaction=True, swap_conn=lb_conn)
             mapping.swap_into_production(no_swap_transaction=True, swap_conn=lb_conn)
             can.swap_into_production(no_swap_transaction=True, swap_conn=lb_conn)
             can_rec_rel.swap_into_production(no_swap_transaction=True, swap_conn=lb_conn)
@@ -155,7 +149,6 @@ def create_canonical_musicbrainz_data(use_lb_conn: bool):
             lb_conn.close()
         else:
             releases.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
-            mapping_release.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
             mapping.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
             can.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
             can_rec_rel.swap_into_production(no_swap_transaction=True, swap_conn=mb_conn)
